+++ src/headers/common.h
@@ -1,3 +1,3 @@
 
 void readKernelSource(const char *filename, char **source);
-double *create2DSpace(int width, int height);
+int *create2DSpace(int width, int height);
